CLEANUP: move thread function declarations to thread.h
